+++ botx/models/responses.py
@@ -44,3 +44,18 @@ class StealthResponse(BaseModel):
     """list of errors"""
     error_data: Optional[Dict[str, str]] = None
     """error data"""
+
+
+class AddRemoveUserResponse(BaseModel):
+    """Entity that will be returned from BotX API as a result of add/remove users"""
+
+    status: Statuses = Statuses.ok
+    """operation status"""
+    result: Optional[bool] = None
+    """operation result."""
+    reason: Optional[str] = None
+    """reason of error."""
+    errors: Optional[List[str]] = None
+    """list of errors"""
+    error_data: Optional[Dict[str, str]] = None
+    """error data"""